// emasaku-cloudflare--semua-menjadi-satu/types.ts
export enum UserRole {
  MEMBER = 'Member',
  ADMIN = 'Admin',
  DEVELOPER = 'Developer',
}

export interface User {
  id: string;
  name: string;
  email?: string;
  avatarUrl: string;
  role: UserRole;
  joinedDate: string;
  whatsapp?: string;
  driverBalance?: number;
  isRetailPartner?: boolean;
  isLaundryPartner?: boolean;
  isSeller?: boolean;
  isWastePartner?: boolean;
}
// ...
export interface Product {
    id: number;
    name: string;
    price: number;
    description: string;
    category: string;
    imageUrl: string;
    stock: number;
    variants?: string[];
    seller: {
        id: string;
        name: string;
        avatarUrl: string;
        location: string;
    };
}
// ...
export type ThemeColors = { [key: string]: string };